pyproject.toml
[project]
name = "boe-dog-summary"
version = "0.1.0"
description = "Daily Telegram bot that summarizes DOG and BOE official bulletins"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
]

[project.scripts]
boe-dog-summary = "tg_summary.main:main"

[tool.setuptools]
package-dir = { "" = "src" }
